clean-up bayespairing
- separates a lot of processing into a separate helper class
